+++ src/main/java/edu/illinois/library/cantaloupe/cache/CompletableNullOutputStream.java
@@ -1,0 +1,14 @@
+package edu.illinois.library.cantaloupe.cache;
+
+import java.io.IOException;
+
+/**
+ * @since 4.1.9
+ */
+public class CompletableNullOutputStream extends CompletableOutputStream {
+
+    @Override
+    public void write(int b) throws IOException {
+    }
+
+}