+++ src/main/java/edu/illinois/library/cantaloupe/cache/RedisCache.java
@@ -21,7 +21,6 @@ import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
-import java.io.OutputStream;
 import java.nio.ByteBuffer;
 import java.nio.charset.StandardCharsets;
 import java.util.Optional;
@@ -137,7 +136,7 @@ class RedisCache implements DerivativeCache {
     /**
      * Buffers written data and then writes it asynchronously to Redis.
      */
-    private static class RedisOutputStream extends OutputStream {
+    private static class RedisOutputStream extends CompletableOutputStream {
 
         private final ByteArrayOutputStream bufferStream = new ByteArrayOutputStream();
         private final StatefulRedisConnection<String, byte[]> connection;
@@ -155,8 +154,10 @@ class RedisCache implements DerivativeCache {
         @Override
         public void close() throws IOException {
             try {
-                connection.async().hset(hashKey, valueKey,
-                        bufferStream.toByteArray());
+                if (isCompletelyWritten()) {
+                    connection.async().hset(hashKey, valueKey,
+                            bufferStream.toByteArray());
+                }
             } finally {
                 super.close();
             }
@@ -240,7 +241,8 @@ class RedisCache implements DerivativeCache {
     }
 
     @Override
-    public OutputStream newDerivativeImageOutputStream(OperationList opList) {
+    public CompletableOutputStream
+    newDerivativeImageOutputStream(OperationList opList) {
         return new RedisOutputStream(IMAGE_HASH_KEY, imageKey(opList),
                 getConnection());
     }