+++ src/test/java/edu/illinois/library/cantaloupe/cache/JdbcCacheTest.java
@@ -18,7 +18,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
-import java.io.OutputStream;
 import java.nio.file.Files;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -104,8 +103,10 @@ public class JdbcCacheTest extends AbstractCacheTest {
         // persist some derivative images
         OperationList ops = new OperationList();
 
-        try (OutputStream os = instance.newDerivativeImageOutputStream(ops)) {
+        try (CompletableOutputStream os =
+                     instance.newDerivativeImageOutputStream(ops)) {
             Files.copy(TestUtil.getImage(IMAGE), os);
+            os.setCompletelyWritten(true);
         }
 
         Crop crop     = new CropByPixels(50, 50, 50, 50);
@@ -115,8 +116,10 @@ public class JdbcCacheTest extends AbstractCacheTest {
                 .withOperations(crop, scale, rotate)
                 .build();
 
-        try (OutputStream os = instance.newDerivativeImageOutputStream(ops)) {
+        try (CompletableOutputStream os =
+                     instance.newDerivativeImageOutputStream(ops)) {
             Files.copy(TestUtil.getImage(IMAGE), os);
+            os.setCompletelyWritten(true);
         }
 
         crop   = new CropByPixels(10, 20, 50, 90);
@@ -126,8 +129,10 @@ public class JdbcCacheTest extends AbstractCacheTest {
                 .withOperations(crop, scale, rotate)
                 .build();
 
-        try (OutputStream os = instance.newDerivativeImageOutputStream(ops)) {
+        try (CompletableOutputStream os =
+                     instance.newDerivativeImageOutputStream(ops)) {
             Files.copy(TestUtil.getImage(IMAGE), os);
+            os.setCompletelyWritten(true);
         }
 
         // persist some infos corresponding to the above images