+++ src/test/java/edu/illinois/library/cantaloupe/cache/MockCache.java
@@ -63,8 +63,8 @@ class MockCache implements DerivativeCache, SourceCache {
     }
 
     @Override
-    public OutputStream newDerivativeImageOutputStream(OperationList opList)
-            throws IOException {
+    public CompletableOutputStream
+    newDerivativeImageOutputStream(OperationList opList) throws IOException {
         return null;
     }