+++ src/test/java/edu/illinois/library/cantaloupe/cache/S3CacheTest.java
@@ -20,7 +20,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import software.amazon.awssdk.services.s3.S3Client;
 
-import java.io.OutputStream;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.nio.file.Files;
@@ -240,8 +239,10 @@ public class S3CacheTest extends AbstractCacheTest {
         // Add an image.
         // N.B.: This method may return before data is fully (or even
         // partially) written to the cache.
-        try (OutputStream os = instance.newDerivativeImageOutputStream(ops)) {
+        try (CompletableOutputStream os =
+                     instance.newDerivativeImageOutputStream(ops)) {
             Files.copy(fixture, os);
+            os.setCompletelyWritten(true);
         }
 
         // Wait for it to finish, hopefully.